requirements.txt
altgraph==0.16.1
bgtunnel==0.4.1
Click==7.0
mailer==0.8.1
mysqlclient==1.3.13
pkg-resources==0.0.0
PyInstaller==3.5
python-dotenv==0.10.1
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
